an invalid address pointer to zero bytes of data is still valid
git-svn-id: svn+ssh://market.scs.stanford.edu/home/u1/asbestos/repo/jos64/trunk@403 927ebb28-5305-0410-b3ca-c4b640df69b7

+++ kern/arch/amd64/pmap.h
@@ -70,7 +70,7 @@ kva2pa (void *kva)
  * Changes *ptrp such that it will not reference a kernel address,
  * and makes sure the address is paged in (might return -E_RESTART).
  */
-int  page_user_incore(void **ptrp, int nbytes);
+int  page_user_incore(void **ptrp, uint64_t nbytes);
 
 #endif /* !__ASSEMBLER__ */
